decluttering GPXMapView
Core Data stuff moved to CoreDataHelper.swift
